+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## Added
-- Support for trim patterns added in "More Armor Trims" v1.1.0
+- Frostiful support.
+
+## Fixed
+- Compatibility with Elytra Trims lmao.